Stage 0: widen PgfCodec DataT from short to int (pgf-all-image-modes.md)
Pre-Stage-1 verification found managed-pgf-codec.md's "DataT is INT16,
not INT32" claim was wrong: PGFplatform.h defines __PGF32SUPPORT__ by
default, and this repo's CMakeLists.txt never defines NPGF32 to turn it
off, so the real oracle build has always used INT32 coefficients and
MaxBitPlanes=31. It was silent for RGBA (8-bit-per-channel data never
approaches Int16's range) but would have broken this PRD's 16-bit-per-
channel groups (Gray16/Lab48/RGB48/CMYK64), where a raw channel value of
0 is exactly -32768 after the YUV offset.

Widens the coefficient type across the shared codec core (wavelet
transform, subbands, macroblocks, entropy coder) to match the real
oracle build, and fixes a related native-shim bug found along the way:
pgf_debug_decode_channel declared an int16_t output buffer but memcpy'd
sizeof(DataT)==4-byte elements into it, a latent heap overflow (dead
code - no test called it, but it's needed for later-stage debugging of
the new 16-bit-channel modes).

Full regression: PictTag.PgfCodec.Tests 920/920, PictTag.Data.Tests
15/15, both green. Adds a boundary-value entropy-coder test
(ValuesBeyondInt16Range) proving the widening buys real range, not just
a type rename.

+++ native/PictTag.PgfDecoder/shim.cpp
@@ -265,7 +265,8 @@ PICTTAG_EXPORT void pgf_close(PgfDecoderHandle* handle)
 // than something only reachable via real digiKam-produced fixtures.
 
 // Encodes a BGRA buffer (top-down, pitch = width*4) into a newly heap-allocated PGF byte buffer.
-// quality: 0 = lossless, up to MaxQuality (15 in this build - see PGFtypes.h; SetHeader's own
+// quality: 0 = lossless, up to MaxQuality (31 in this build - __PGF32SUPPORT__ is active by
+// PGFplatform.h's own default, since NPGF32 is never defined - see PGFtypes.h; SetHeader's own
 // ASSERT(header.quality <= MaxQuality) is the actual enforced bound, not the narrower "0/4/6" preset
 // list PGFHeader's doc comment calls out by name).
 // The caller owns the returned buffer and must free it with pgf_free_encoded - an owned-pointer
@@ -383,7 +384,7 @@ PICTTAG_EXPORT void pgf_free_encoded(uint8_t* data)
 // twice per test process until this is properly root-caused.
 PICTTAG_EXPORT bool pgf_debug_decode_channel(
     const uint8_t* data, size_t dataLen, int level, int channel,
-    int16_t* outBuffer, size_t outBufferLen, uint32_t* outWidth, uint32_t* outHeight)
+    int32_t* outBuffer, size_t outBufferLen, uint32_t* outWidth, uint32_t* outHeight)
 {
     if (data == nullptr || dataLen == 0 || outBuffer == nullptr || outWidth == nullptr ||
         outHeight == nullptr || level < 0 || channel < 0)
